Features 27,28,29,30 added

+++ src/SpecialLinkLoginPages.php
@@ -13,9 +13,7 @@ class SpecialLinkLoginPages extends SpecialPage {
 
 
 	function execute( $par ) {
-		$this->checkPermissions('linklogin-link');
 		$this->setHeaders();
-		
 		LinkLogin::populateLoginTokens();
 
 		//Show Users and Pages belonging to the group set in $par
@@ -42,6 +40,8 @@ class SpecialLinkLoginPages extends SpecialPage {
 		$output->addHTML('<table class="table table-bordered table-sm"><tr>');
 		$output->addHTML('<th>' . wfMessage("linklogin-categories")->text() . '</th>');
 		$output->addHTML('</tr>');
+		$categories = array_unique($categories);
+		natcasesort($categories);
 		foreach( $categories as $category ) {
 			$output->addHTML('<tr>');
 			$url = SpecialPage::getTitleFor( 'LinkLoginPages' )->getLocalURL() . '/' . $category;
@@ -61,6 +61,12 @@ class SpecialLinkLoginPages extends SpecialPage {
 		$output = $this->getOutput();
 		$output->addModules("ext.linklogin-mapping");
 		$output->addWikiTextAsInterface('{{#tweekihide:sidebar-right}}');
+		$api_access = false;
+		if( MediaWikiServices::getInstance()
+			->getPermissionManager()
+			->userHasRight($this->getUser(),'linklogin-link') ) {
+				$api_access = true;
+		} 
 
 		$groups = LinkLogin::getLinkLoginGroupsByCategory($par);
 
@@ -80,6 +86,7 @@ class SpecialLinkLoginPages extends SpecialPage {
 		$result = SMWQueryProcessor::getResultFromQuery( $query, $processed_params, SMW_OUTPUT_WIKI, SMWQueryProcessor::SPECIAL_PAGE );
 		$pages = explode( '<SEP>', $result );
 		$displaytitles = [];
+		$displaytitle = '';
 		foreach( $pages as $page ) {
 			if( !empty($page) ) {
 				list( $title, $displaytitle ) = explode("<PROP>", $page );
@@ -122,6 +129,16 @@ class SpecialLinkLoginPages extends SpecialPage {
 			return strnatcasecmp( $a->displaytitle, $b->displaytitle );
 		} );
 
+		$assoc_groups = LinkLogin::getLinkLoginGroupsByCategory($par);
+
+		$output->setPageTitle( $this->getDescription() . ' (' . wfMessage('linklogin-category') . ': '  . $par . ')');
+		$output->addHTML('<div class="col text-center"><a href="' . SpecialPage::getTitleFor( 'LinkLoginPages' )->getLocalURL() . '"><button type="button" class="btn btn-secondary translate-middle">' . wfMessage('linklogin-overview') . '</button></a></div>');
+		$output->addHTML('<div class="col" style="margin: 10px 0px">' . wfMessage("linklogin-associated") . ' ' . wfMessage("linklogin-groups"). ': ');
+		foreach($assoc_groups as $assoc_group){
+			$url = SpecialPage::getTitleFor( 'LinkLoginUsers' )->getLocalURL() . '/' . $assoc_group;
+			$output->addHTML('<a href="' . $url . '">' . $assoc_group . '</a>' . ' ');
+		}
+		$output->addHTML('</div>');
 		$output->addHTML('<container id="linklogin-body">');
 		$output->addHTML('<table class="table table-bordered table-sm"><tr>');
 		$output->addHTML('<th>' . wfMessage("linklogin-page")->text() . '</th>');
@@ -152,8 +169,10 @@ class SpecialLinkLoginPages extends SpecialPage {
 			if( !empty( $user ) ) {
 				$output->addHTML('<td id=' . $page->id . 'User' . '>');
 				$output->addHTML('<span>' . $user . '</span>' . " ");
-				$output->addHTML('<a href="#"><i class="fa fa-pen edit" title="' . wfMessage('linklogin-edit-user') . '" data-toggle="tooltip"></i></a>');
-				$output->addHTML('<a href="#" class="unlink users ml-2"><i class="fa fa-times" title="' . wfMessage('linklogin-unlink') . '" data-toggle="tooltip"></i></a>');
+				if( $api_access ) {
+					$output->addHTML('<a href="#"><i class="fa fa-pen edit" title="' . wfMessage('linklogin-edit-user') . '" data-toggle="tooltip"></i></a>');
+					$output->addHTML('<a href="#" class="unlink users ml-2"><i class="fa fa-times" title="' . wfMessage('linklogin-unlink') . '" data-toggle="tooltip"></i></a>');
+				}
 				$output->addHTML('</td>');
 			} else {             
 				foreach( $groups as $group ) {
@@ -176,22 +195,30 @@ class SpecialLinkLoginPages extends SpecialPage {
 				//User Column
 				$output->addHTML('<td id=' . $page->id . 'User' . '>');
 				$output->addHTML('<container id='. $page->id . 'Fragment>');
-				$output->addHTML('<div class="dropdown">');
-				$output->addHTML('<a class="dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">');
-				$output->addHTML(wfMessage("linklogin-assign-user")->text());
-				$output->addHTML('</a>');
-				$output->addHTML('<div class="dropdown-menu" aria-labelledby="dropdownMenuButton">');
-				foreach( $users as $user ){
-					$output->addHTML('<a href="#" class="dropdown-item user">' . $user . '</a>');
+				if( $api_access ) {
+					$output->addHTML('<div class="dropdown">');
+					$output->addHTML('<a class="dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">');
+					$output->addHTML(wfMessage("linklogin-assign-user")->text());
+					$output->addHTML('</a>');
+					$output->addHTML('<div class="dropdown-menu userlist" aria-labelledby="dropdownMenuButton">');
+					foreach( $users as $user ){
+						$output->addHTML('<a href="#" class="dropdown-item user">' . $user . '</a>');
+					}
+					$output->addHTML('</div>');
 				}
-				$output->addHTML('</div>');
 
-				//Neuen User anlegen
-				$output->addHTML('<form class="user-create form-inline" novalidate>');
-				$output->addHTML('<input id="' . $page->id .'Inputfield" class="username md-textarea form-control mr-1" rows="1" style="width:200px" placeholder="' . wfMessage("linklogin-user-create-placeholder")->text() . '">');
-				$output->addHTML('<button type="submit" class="btn btn-primary create">' . wfMessage("linklogin-user-create-short")->text() . '</button>');
-				$output->addHTML('<small id="' . $page->id . 'userError" class="userError text-danger"></small>');
-				$output->addHTML('</form>');
+				//Neuen User anlegen, wenn User 'createaccount' Rechte besitzt
+				if( MediaWikiServices::getInstance()
+					->getPermissionManager()
+					->userHasRight($this->getUser(),
+					'createaccount') ) 
+				{
+					$output->addHTML('<form class="user-create form-inline" novalidate>');
+					$output->addHTML('<input id="' . $page->id .'Inputfield" class="username md-textarea form-control mr-1" rows="1" style="width:200px" placeholder="' . wfMessage("linklogin-user-create-placeholder")->text() . '">');
+					$output->addHTML('<button type="submit" class="btn btn-primary create">' . wfMessage("linklogin-user-create-short")->text() . '</button>');
+					$output->addHTML('<small id="' . $page->id . 'userError" class="userError text-danger"></small>');
+					$output->addHTML('</form>');
+				}
 				$output->addHTML('</td>');
 				$output->addHTML('</container>');
 				$output->addHTML('</tr>');
