+++ src/SpecialLinkLoginUsers.php
@@ -13,7 +13,6 @@ class SpecialLinkLoginUsers extends SpecialPage {
 
 
 	function execute( $par ) {
-		$this->checkPermissions('linklogin-link');
 		$this->setHeaders();
 
 		LinkLogin::populateLoginTokens();
@@ -64,6 +63,13 @@ class SpecialLinkLoginUsers extends SpecialPage {
 		$output = $this->getOutput();
 		$output->addModules("ext.linklogin-mapping");
 		$output->addWikiTextAsInterface('{{#tweekihide:sidebar-right}}');
+
+		$api_access = false;
+		if( MediaWikiServices::getInstance()
+			->getPermissionManager()
+			->userHasRight($this->getUser(), 'linklogin-link') ) {
+				$api_access = true;
+		} 
 
 		$uom = MediaWikiServices::getInstance()->getUserOptionsManager();
 
@@ -151,7 +157,7 @@ class SpecialLinkLoginUsers extends SpecialPage {
 			]
 		) ?: [];
 		if( !empty($pages) ) {
-			foreach($pages as $page){
+			foreach( $pages as $page ) {
 				$page->displaytitle = array_search( str_replace( '_' ,' ', $page->page_title ), $displaytitles );
 				$unlinked_pages[$page->page_id] = $page->displaytitle;
 			}
@@ -163,6 +169,21 @@ class SpecialLinkLoginUsers extends SpecialPage {
 			}
 		}
 		natcasesort($unlinked_pages);
+
+		$assoc_categories = LinkLogin::getLinkLoginCategories([$par]);
+
+		if( wfMessage( 'group-' . $par )->exists() ) {
+			$par = wfMessage( 'group-' . $par )->text();
+		}
+		$output->setPageTitle( $this->getDescription() . ' (' . wfMessage('linklogin-group') . ': '  . $par . ')');
+
+		$output->addHTML('<div class="col text-center"><a href="' . SpecialPage::getTitleFor( 'LinkLoginUsers' )->getLocalURL() . '"><button type="button" class="btn btn-secondary translate-middle">' . wfMessage('linklogin-overview') . '</button></a></div>');
+		$output->addHTML('<div class="col" style="margin: 10px 0px">' . wfMessage("linklogin-associated") . ' ' . wfMessage("linklogin-categories"). ': ');
+		foreach($assoc_categories as $assoc_cat){
+			$url = SpecialPage::getTitleFor( 'LinkLoginPages' )->getLocalURL() . '/' . $assoc_cat;
+			$output->addHTML('<a href="' . $url . '">' . $assoc_cat . '</a>' . ' ');
+		}
+		$output->addHTML('</div>');
 		$output->addHTML('<container id="linklogin-body">');
 		$output->addHTML('<table class="table table-bordered table-sm"><tr>');
 		$output->addHTML('<th>' . wfMessage("linklogin-username")->text() . '</th>');
@@ -178,31 +199,38 @@ class SpecialLinkLoginUsers extends SpecialPage {
 		foreach( $users as $user ) {
 			$user_name = str_replace(' ', '_', $user->user_name);
 			$output->addHTML('<tr id=' . '"' . $user_name . '"' . '>');
-			$output->addHTML('<td>' . '<span>' . $user->user_name . '</span>' . ' ' . '<a href="#"><i class="fa fa-pen edit" title="' . wfMessage('linklogin-edit-user') . '" data-toggle="tooltip"></i></a>' . '</td>');
+			if( $api_access ) {
+				$output->addHTML('<td>' . '<span>' . $user->user_name . '</span>' . ' ' . '<a href="#"><i class="fa fa-pen edit" title="' . wfMessage('linklogin-edit-user') . '" data-toggle="tooltip"></i></a>' . '</td>');
+			} else {
+				$output->addHTML('<td>' . '<span>' . $user->user_name . '</span>' . '</td>');
+			}
 			$output->addHTML('<td id="' . $user_name . 'Pages">');
 			if( array_key_exists($user->user_name, $linked_pages)){
 				$output->addHTML('<ul id="' . $user_name . 'List">');
 				foreach( $linked_pages[$user->user_name] as $id_key => $linked_page){
 					$output->addHTML('<li id="listitem-' . $id_key . '">');
 					$output->addHTML('<span>' . $linked_page . '</span>');
-					$output->addHTML('<a href="#" class="unlink pages ml-2"><i class="fa fa-times" title="' . wfMessage('linklogin-unlink') . '" data-toggle="tooltip"></i></a>');
+					if( $api_access ) {
+						$output->addHTML('<a href="#" class="unlink pages ml-2"><i class="fa fa-times" title="' . wfMessage('linklogin-unlink') . '" data-toggle="tooltip"></i></a>');
+					}
 					$output->addHTML('</li>');
 				}
 				$output->addHTML('</ul>');
 			}
-			$output->addHTML('<div class="dropdown">');
-			$output->addHTML('<a class="dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">');
-			$output->addHTML(wfMessage('linklogin-assign-page')->text());
-			$output->addHTML('</a>');
-			$output->addHTML('<div class="dropdown-menu pageslist" aria-labelledby="dropdownMenuButton">');
-			foreach($unlinked_pages as $key => $unlinked_page){
-
-				// show only pages not already associated with the user
-				if(!in_array($unlinked_page,$linked_pages)){
-					$output->addHTML('<a href="#" class="dropdown-item pages" id="dropdownitem-'. $key .'">' . $unlinked_page . '</a>');
+			if( $api_access ) {
+				$output->addHTML('<div class="dropdown">');
+				$output->addHTML('<a class="dropdown-toggle" type="button" id="dropdownMenuButton" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">');
+				$output->addHTML(wfMessage('linklogin-assign-page')->text());
+				$output->addHTML('</a>');
+				$output->addHTML('<div class="dropdown-menu pageslist" aria-labelledby="dropdownMenuButton">');
+				foreach($unlinked_pages as $key => $unlinked_page){
+					// show only pages not already associated with the user
+					if(!in_array($unlinked_page,$linked_pages)){
+						$output->addHTML('<a href="#" class="dropdown-item pages" id="dropdownitem-'. $key .'">' . $unlinked_page . '</a>');
+					}
 				}
+				$output->addHTML('</div></div>');
 			}
-			$output->addHTML('</div></div>');
 			$output->addHTML('</td>');
 
 			//Look if User has an e-mail assoiciated to them
@@ -218,12 +246,14 @@ class SpecialLinkLoginUsers extends SpecialPage {
 
 			//Add quick custom mail icons 
 			$output->addHTML('<td class="semorg-showedit">');
-			if( !is_null($loginpage) &&  !is_null($user->user_email_token)) {
-				$link = $this->createCustomMailLink($loginpage,$user);
-				$output->addHTML('<a id="' . $link . '" class="copy clipboard mr-2" href="#" title="' . wfMessage('linklogin-clipboard')->text() . '" data-toggle="tooltip"><i class="fa fa-clipboard"></i></a>');
-				if( !empty($email) ){
-					$encoded_link = urlencode($link);
-					$output->addHTML('<a href="mailto:' . $email .'?body=' . $encoded_link . '"><i class="fa fa-envelope fa-sm" data-toggle="tooltip" title="' . wfMessage('linklogin-mail-link')->text() . '"></i></a>');
+			if( $api_access ){
+				if( !is_null($loginpage) &&  !is_null($user->user_email_token)) {
+					$link = $this->createCustomMailLink($loginpage,$user);
+					$output->addHTML('<a id="' . $link . '" class="copy clipboard mr-2" href="#" title="' . wfMessage('linklogin-clipboard')->text() . '" data-toggle="tooltip"><i class="fa fa-clipboard"></i></a>');
+					if( !empty($email) ){
+						$encoded_link = urlencode($link);
+						$output->addHTML('<a href="mailto:' . $email .'?body=' . $encoded_link . '"><i class="fa fa-envelope fa-sm" data-toggle="tooltip" title="' . wfMessage('linklogin-mail-link')->text() . '"></i></a>');
+					}
 				}
 			}
 			$output->addHTML('</td>');
